MSVC fixes + clean out tree
git-svn-id: https://svn.eduke32.com/eduke32@1399 1a8010ca-5511-0410-912e-c29ae57300e0

+++ polymer/eduke32/source/menus.c
@@ -3718,7 +3718,7 @@ cheat_for_port_credits:
         mgametextpal(40,118+9+9+9+9,"Advanced mouse setup",MENUHIGHLIGHT((MAXMOUSEBUTTONS-2)*2+2+2+2),10);
 
         {
-            int32_t sense = CONTROL_MouseSensitivity * 2;
+            int32_t sense = (int32_t)(CONTROL_MouseSensitivity * 2.0f);
             barsm(248,126,&sense,2,x==(MAXMOUSEBUTTONS-2)*2+2,MENUHIGHLIGHT((MAXMOUSEBUTTONS-2)*2+2),PHX(-7));
             CONTROL_MouseSensitivity = sense / 2.0f;
         }
